+++ changelog.md
@@ -32,6 +32,8 @@
 - Added deny and success action for each requirement. In addition to its global actions, you can add them for each requirement. Perfect to set the error message for the player.
 - When creating ItemStacks, the plugin will no longer use PlaceholderAPI if the item contains no placeholders.
 - Added cache system for ItemStack. If ItemStack does not use placeholder API, then it will be created once and cached. All inventories will therefore use the same ItemStack. This allows to improve performance (from a few thousand nano seconds per item, so if you have a lot of items in your inventory the gain can be very important).
+- Fixed command /zm open. The opening arguments were only taken into account if the command had more than 5 arguments instead of starting from 5.
+- Fixed command /zm reload with commands.
 
 # 1.0.2.0